refactored Gallery
