+++ src/components/Gallery/index.ts
@@ -1,3 +1,3 @@
-export { default as Gallery } from "./Gallery";
-export { default as useGallery } from "./useGallery";
+export * from "./Gallery";
+export * from "./useGallery";
 export * from "./Gallery.types";
