+++ src/components/GalleryItem/GalleryItemContext.ts
@@ -1,5 +1,5 @@
 import { createContext } from "react";
-import { GALLERY_ITEM_NO_PROVIDER_FLAG, GalleryItemState } from "./GalleryItem.types";
+import { GALLERY_ITEM_NO_PROVIDER_FLAG, GalleryItemState } from ".";
 
 const GalleryItemContext = createContext<GalleryItemState | typeof GALLERY_ITEM_NO_PROVIDER_FLAG>(
   GALLERY_ITEM_NO_PROVIDER_FLAG,
